Переделал requests.py в два класса GET и POST перенес функцию декодирования в main
