+++ wsgi_framework/main.py
@@ -1,8 +1,11 @@
-from wsgi_framework.requests import Requests
+from quopri import decodestring
+
+from wsgi_framework.requests import GetRequests, PostRequests
 from wsgi_framework.templator import render
 
 
 class Framework:
+    """Класс Framework - основа фреймворка"""
     def __init__(self, routes_obj_ls, fronts_obj_ls):
         self.lsRoutes = routes_obj_ls
         self.lsFronts = fronts_obj_ls
@@ -20,21 +23,24 @@ class Framework:
         method = environ['REQUEST_METHOD']
         request[method] = method
         if method == 'GET':
-            request_params = Requests().get_Get_request_params(environ)
-            request['request_params'] = request_params
-            print(f'Получены GET-параметры: {request_params}')
+            request_params = GetRequests().get_request_params(environ)
+            request['request_params'] = Framework.decode_value(request_params)
+            print(f'Получены GET-параметры: {Framework.decode_value(request_params)}')
 
         if method =='POST':
-            data = Requests().get_Post_request_params(environ)
-            request['data'] = data
-            print(f'Получен POST-запрос: {data}')
+            data = PostRequests().get_request_params(environ)
+            request['data'] = Framework.decode_value(data)
+            print(f'Получен POST-запрос: {Framework.decode_value(data)}')
 
+        # находим нужный контроллер
         # отработка паттерна Page Controller
         if path in self.lsRoutes:
             view = self.lsRoutes[path]
         else:
             view = PageNotFound404()
 
+        # наполняем словарь request элементами
+        # этот словарь получат все контроллеры
         # Отработка паттерна Front Controller
         for front in self.lsFronts:
             front(request)
@@ -44,14 +50,15 @@ class Framework:
         start_response(code, [('Content-Type', 'text/html')])
         return [body.encode('utf-8')]
 
-    # def parse_input_data(self, data: str):
-    #     result = {}
-    #     if data:
-    #         params = data.split('&')
-    #         for item in params:
-    #             key, value = item.split('=')
-    #             result[key] = value
-    #     return result
+    # Декодер данных
+    @staticmethod
+    def decode_value(data):
+        new_data = {}
+        for key, value in data.items():
+            val = bytes(value.replace("%", "=").replace("+", " "), 'UTF-8')
+            val_decode_str = decodestring(val).decode('UTF-8')
+            new_data[key] = val_decode_str
+        return new_data
 
 
 class PageNotFound404:
