+++ wsgi_framework/requests.py
@@ -1,7 +1,4 @@
-from quopri import decodestring
-
-
-class Requests:
+class GetRequests:
     # Парсер данных
     @staticmethod
     def parse_input_data(data: str):
@@ -13,28 +10,25 @@ class Requests:
                 result[key] = value
         return result
 
-    # Декодер данных
-    @staticmethod
-    def decode_value(data):
-        new_data = {}
-        for key, value in data.items():
-            val = bytes(value.replace("%", "=").replace("+", " "), 'UTF-8')
-            val_decode_str = decodestring(val).decode('UTF-8')
-            new_data[key] = val_decode_str
-        return new_data
-
     # Метод получения параметров GET запроса
     @staticmethod
-    def get_Get_request_params(environ):
+    def get_request_params(environ):
         query_string = environ['QUERY_STRING']
-        request_params = Requests.parse_input_data(query_string)
-        return Requests.decode_value(request_params)
+        request_params = GetRequests.parse_input_data(query_string)
+        return request_params
 
-    # Метод получения данных POST запроса
-    def get_Post_request_params(self, environ):
-        data = self.get_wsgi_input_data(environ)
-        data = self.parse_wsgi_input_data(data)
-        return Requests.decode_value(data)
+
+class PostRequests:
+    # Парсер данных
+    @staticmethod
+    def parse_input_data(data: str):
+        result = {}
+        if data:
+            params = data.split('&')
+            for item in params:
+                key, value = item.split('=')
+                result[key] = value
+        return result
 
     # Получаем данные POST запроса в bytes
     @staticmethod
@@ -49,8 +43,17 @@ class Requests:
     def parse_wsgi_input_data(self, data: bytes) -> dict:
         result = {}
         if data:
+            # декодируем данные
             data_str = data.decode(encoding="utf-8")
             print(f'Строка данных после декодирования: {data_str}')
-            print(type(data_str))
+            # собираем их в словарь
             result = self.parse_input_data(data_str)
         return result
+
+    # Метод получения данных POST запроса
+    def get_request_params(self, environ):
+        # получаем данные
+        data = self.get_wsgi_input_data(environ)
+        # превращаем данные в словарь
+        data = self.parse_wsgi_input_data(data)
+        return data
